Deep clean up.

+++ ManuvrOS/DataStructures/Argument.cpp
@@ -685,7 +685,7 @@ void Argument::valToString(StringBuilder* out) {
       out->concatf("%u", (uintptr_t) pointer());
       break;
     case TCode::FLOAT:
-      out->concatf("%.4f", *((float*) &buf));
+      out->concatf("%.4f", (double) *((float*) &buf));
       break;
     case TCode::BOOLEAN:
       out->concatf("%s", ((uintptr_t) pointer() ? "true" : "false"));
